+++ x/ibc/07-tendermint/types/msgs.go
@@ -23,9 +23,9 @@ const (
 )
 
 var (
-	_ clientexported.MsgCreateClient     = MsgCreateClient{}
-	_ clientexported.MsgUpdateClient     = MsgUpdateClient{}
-	_ evidenceexported.MsgSubmitEvidence = MsgSubmitClientMisbehaviour{}
+	_ clientexported.MsgCreateClient     = &MsgCreateClient{}
+	_ clientexported.MsgUpdateClient     = &MsgUpdateClient{}
+	_ evidenceexported.MsgSubmitEvidence = &MsgSubmitClientMisbehaviour{}
 )
 
 // MsgCreateClient defines a message to create an IBC client
@@ -38,6 +38,14 @@ type MsgCreateClient struct {
 	MaxClockDrift   time.Duration   `json:"max_clock_drift" yaml:"max_clock_drift"`
 	Signer          sdk.AccAddress  `json:"address" yaml:"address"`
 }
+
+// this is a constant to satisfy the linter
+const TODO = "TODO"
+
+// dummy implementation of proto.Message
+func (msg MsgCreateClient) Reset()         {}
+func (msg MsgCreateClient) String() string { return TODO }
+func (msg MsgCreateClient) ProtoMessage()  {}
 
 // NewMsgCreateClient creates a new MsgCreateClient instance
 func NewMsgCreateClient(
@@ -129,6 +137,11 @@ type MsgUpdateClient struct {
 	Signer   sdk.AccAddress `json:"address" yaml:"address"`
 }
 
+// dummy implementation of proto.Message
+func (msg MsgUpdateClient) Reset()         {}
+func (msg MsgUpdateClient) String() string { return TODO }
+func (msg MsgUpdateClient) ProtoMessage()  {}
+
 // NewMsgUpdateClient creates a new MsgUpdateClient instance
 func NewMsgUpdateClient(id string, header Header, signer sdk.AccAddress) MsgUpdateClient {
 	return MsgUpdateClient{
@@ -182,6 +195,11 @@ type MsgSubmitClientMisbehaviour struct {
 	Evidence  evidenceexported.Evidence `json:"evidence" yaml:"evidence"`
 	Submitter sdk.AccAddress            `json:"submitter" yaml:"submitter"`
 }
+
+// dummy implementation of proto.Message
+func (msg MsgSubmitClientMisbehaviour) Reset()         {}
+func (msg MsgSubmitClientMisbehaviour) String() string { return TODO }
+func (msg MsgSubmitClientMisbehaviour) ProtoMessage()  {}
 
 // NewMsgSubmitClientMisbehaviour creates a new MsgSubmitClientMisbehaviour
 // instance.